Extract Google Auth QR parser
